make UnitRegistry.__contains__ a bit smarter
