+++ unyt/unit_object.py
@@ -64,14 +64,11 @@ from unyt.exceptions import (
     UnitsNotReducible,
 )
 from unyt._physical_ratios import speed_of_light_cm_per_s
-from unyt._unit_lookup_table import (
-    unit_prefixes,
-    latex_prefixes,
-)
 from unyt.unit_registry import (
     default_unit_registry,
+    _lookup_unit_symbol,
     UnitRegistry,
-    UnitParseError
+    UnitParseError,
 )
 
 sympy_one = sympify(1)
@@ -896,85 +893,6 @@ def _get_unit_data_from_expr(unit_expr, unit_symbol_lut):
     raise UnitParseError("Cannot parse for unit data from '%s'. Please supply"
                          " an expression of only Unit, Symbol, Pow, and Mul"
                          "objects." % str(unit_expr))
-
-
-def _lookup_unit_symbol(symbol_str, unit_symbol_lut):
-    """
-    Searches for the unit data tuple corresponding to the given symbol.
-
-    Parameters
-    ----------
-    symbol_str : str
-        The unit symbol to look up.
-    unit_symbol_lut : dict
-        Dictionary with symbols as keys and unit data tuples as values.
-
-    """
-    if symbol_str in unit_symbol_lut:
-        # lookup successful, return the tuple directly
-        return unit_symbol_lut[symbol_str]
-
-    # could still be a known symbol with a prefix
-    possible_prefix = symbol_str[0]
-
-    if symbol_str[:2] == 'da':
-        possible_prefix = 'da'
-
-    if possible_prefix in unit_prefixes:
-        # the first character could be a prefix, check the rest of the symbol
-        symbol_wo_pref = symbol_str[1:]
-
-        # deca is the only prefix with length 2
-        if symbol_str[:2] == 'da':
-            symbol_wo_pref = symbol_str[2:]
-            possible_prefix = 'da'
-
-        prefixable_units = [u for u in unit_symbol_lut
-                            if unit_symbol_lut[u][4]]
-
-        unit_is_si_prefixable = (symbol_wo_pref in unit_symbol_lut and
-                                 symbol_wo_pref in prefixable_units)
-
-        if unit_is_si_prefixable is True:
-            # lookup successful, it's a symbol with a prefix
-            unit_data = unit_symbol_lut[symbol_wo_pref]
-            prefix_value = unit_prefixes[possible_prefix]
-
-            if possible_prefix in latex_prefixes:
-                latex_repr = symbol_str.replace(
-                    possible_prefix, '{'+latex_prefixes[possible_prefix]+'}')
-            else:
-                # Need to add some special handling for comoving units
-                # this is fine for now, but it wouldn't work for a general
-                # unit that has an arbitrary LaTeX representation
-                if symbol_wo_pref != 'cm' and symbol_wo_pref.endswith('cm'):
-                    sub_symbol_wo_prefix = symbol_wo_pref[:-2]
-                    sub_symbol_str = symbol_str[:-2]
-                else:
-                    sub_symbol_wo_prefix = symbol_wo_pref
-                    sub_symbol_str = symbol_str
-                latex_repr = unit_data[3].replace(
-                    '{' + sub_symbol_wo_prefix + '}',
-                    '{' + sub_symbol_str + '}')
-
-            # Leave offset and dimensions the same, but adjust scale factor and
-            # LaTeX representation
-            ret = (unit_data[0] * prefix_value, unit_data[1], unit_data[2],
-                   latex_repr, False)
-
-            unit_symbol_lut[symbol_str] = ret
-
-            return ret
-
-    # no dice
-    if symbol_str.startswith('code_'):
-        raise UnitParseError(
-            "Code units have not been defined. \n"
-            "Try creating the array or quantity using ds.arr or ds.quan "
-            "instead.")
-    else:
-        raise UnitParseError("Could not find unit symbol '%s' in the provided "
-                             "symbols." % symbol_str)
 
 
 def _validate_dimensions(dimensions):
